+++ src/json-rpc/src/DataFormatterFactory.php
@@ -13,6 +13,7 @@ namespace Hyperf\JsonRpc;
 
 use Hyperf\Contract\NormalizerInterface;
 use Hyperf\Rpc\Context;
+use Hyperf\Utils\Serializer\Serializer;
 use Hyperf\Utils\Serializer\SymfonyNormalizer;
 use Psr\Container\ContainerInterface;
 
@@ -23,7 +24,7 @@ class DataFormatterFactory
         /** @var NormalizerInterface $normalizer */
         $normalizer = $container->get(NormalizerInterface::class);
         $context = $container->get(Context::class);
-        if ($normalizer instanceof SymfonyNormalizer) {
+        if ($normalizer instanceof SymfonyNormalizer || $normalizer instanceof Serializer) {
             return new NormalizeDataFormatter($normalizer, $context);
         }
         return new DataFormatter($context);